requirements.txt
open-tamil 0.65
feedparser 5.2.1
matplotlib 1.4
numpy 1.12.0